# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PM Confessional API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /regrets:
    get:
      operationId: listRegrets
      tags: [regrets]
      summary: List regrets with optional filters
      parameters:
        - name: topic_tag
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: guest_name
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of regrets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRegretsResponse"
  /regrets/search:
    post:
      operationId: searchRegrets
      tags: [regrets]
      summary: Semantic search for relevant regrets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRegretsBody"
      responses:
        "200":
          description: Matched regrets ranked by relevance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchRegretsResponse"
...
  /regrets/{id}:
    get:
      operationId: getRegret
      tags: [regrets]
      summary: Get a single regret by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A regret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regret"
        "404":
          description: Not found
...
components:
  schemas:
...
    Regret:
      type: object
      properties:
        id:
          type: integer
        guest_name:
          type: string
        episode_title:
          type: string
        episode_date:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        stage:
          type: string
          enum: [early, growth, scale, general, unknown]
          description: "`unknown` is legacy and being phased out — new rows use `general`."
        topic_tag:
          type: string
          enum: [hiring, pricing, product, growth, culture, fundraising, timing, customers, other]
        regret_statement:
          type: string
        source_quote:
          type: string
        headline_evidence:
          type: ["string", "null"]
          description: Verbatim 8-40 word span from the transcript that proves the headline. Strictly more accurate than `source_quote` (which is the first 500 chars of an 800-word window).
        episode_url:
          type: ["string", "null"]
        relevance_score:
          type: ["number", "null"]
        created_at:
          type: string
      required:
        - id
        - guest_name
        - episode_title
        - stage
        - topic_tag
        - regret_statement
        - source_quote
        - created_at
    ListRegretsResponse:
      type: object
      properties:
        regrets:
          type: array
          items:
            $ref: "#/components/schemas/Regret"
        total:
          type: integer
      required:
        - regrets
        - total
    SearchRegretsBody:
      type: object
      properties:
        query:
          type: string
        limit:
          type: integer
          default: 8
      required:
        - query
    SearchRegretsResponse:
      type: object
      properties:
        regrets:
          type: array
          items:
            $ref: "#/components/schemas/Regret"
        query:
          type: string
        match_count:
          type: integer
          description: Number of regrets above the rerank confidence threshold
        is_fallback:
          type: boolean
          description: True when the Gemini reranker was unavailable and pure cosine ranking was returned instead
        retrieval_mode:
          type: string
          enum: [vector_rerank, vector_only, keyword]
          description: Which pipeline produced the results
      required:
        - regrets
        - query
        - match_count
        - is_fallback
        - retrieval_mode
...